Remove porcelain dependency
The library doesn't use any advanced features of porcelain, which means
the calls can be replaced by regular System.cmd/3 calls.

Additionally, this takes advantage of the fact that `pkg_config` accepts
a list of libraries and not just a single one, so the config string for
all the libraries can be retrieved in just one call instead of library
by library.